Move request shape tests to feature folders
